src/polymetrie/requirements.txt
Flask==3.0.0
requests==2.31.0
psycopg2==2.9.9
redis==5.0.1
python-dotenv==1.0.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.23.0
